:package: feat: Add NewExpense widget for expense input
This commit introduces the `NewExpense` widget, enabling users to input new expenses.
It includes a text field for the expense title and a save button.

Additionally, the `Expenses` widget is updated to display `NewExpense` in a modal bottom sheet when the add button in the `AppBar` is pressed.

Took 26 minutes